# Conflicts:
#	src/main/java/com/anabada/neighbor/club/controller/ClubController.java
#	src/main/java/com/anabada/neighbor/club/service/ClubServiceImpl.java

+++ src/main/java/com/anabada/neighbor/club/controller/ClubController.java
@@ -4,15 +4,14 @@ import com.anabada.neighbor.chat.domain.Chat;
 import com.anabada.neighbor.chat.service.ChattingService;
 import com.anabada.neighbor.club.domain.ClubRequest;
 import com.anabada.neighbor.club.domain.ClubResponse;
-import com.anabada.neighbor.club.domain.ImageRequest;
-import com.anabada.neighbor.club.domain.ImageResponse;
+import com.anabada.neighbor.file.domain.FileRequest;
+import com.anabada.neighbor.file.domain.FileResponse;
 import com.anabada.neighbor.club.domain.entity.Club;
 import com.anabada.neighbor.club.service.ClubService;
-import com.anabada.neighbor.club.service.ImageUtils;
+import com.anabada.neighbor.file.service.FileService;
+import com.anabada.neighbor.file.service.FileUtils;
 import com.anabada.neighbor.config.auth.PrincipalDetails;
-import com.anabada.neighbor.file.controller.ImageController;
-import com.anabada.neighbor.file.domain.ImageInfo;
-import com.anabada.neighbor.file.service.FilesStorageService;
+import com.anabada.neighbor.file.domain.FileInfo;
 import com.anabada.neighbor.used.domain.Post;
 import com.anabada.neighbor.used.service.UsedService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +19,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.method.annotation.MvcUriComponentsBuilder;
 
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
@@ -33,17 +31,18 @@ import java.util.List;
 public class ClubController {
 
     private final ClubService clubService;
-    private final ImageUtils imageUtils;
-    private final FilesStorageService storageService;
     private final ChattingService chattingService;
     private final UsedService usedService;
+    private final FileUtils fileUtils;
+    private final FileService fileService;
 
     public ClubController(ClubService clubService, ImageUtils imageUtils, FilesStorageService storageService, ChattingService chattingService, UsedService usedService) {
+    public ClubController(ClubService clubService, ChattingService chattingService, FileUtils fileUtils, FileService fileService) {
         this.clubService = clubService;
-        this.imageUtils = imageUtils;
-        this.storageService = storageService;
         this.chattingService = chattingService;
         this.usedService = usedService;
+        this.fileUtils = fileUtils;
+        this.fileService = fileService;
     }
 
     @GetMapping("/clubList")
@@ -75,6 +74,7 @@ public class ClubController {
 
     @PostMapping("/clubSave")
     public String clubSave(ClubRequest clubRequest, Model model, @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        // 게시글 객체에 담기
         Post post = Post.builder()
                 .memberId(principalDetails.getMember().getMemberId())
                 .title(clubRequest.getTitle())
@@ -82,20 +82,24 @@ public class ClubController {
                 .postType("club")
                 .build();
         long postId = clubService.savePost(post);
+        //게시물 작성 실패시 postId가 -1로 반환되어 작성실패 메세지 리턴
         if (postId == -1) {
             model.addAttribute("result", "글 등록실패!");
             return "club/clubSave";
         }
+        //게시물 작성 성공시 모임 등록
         Club club = Club.builder()
                 .postId(postId)
                 .memberId(post.getMemberId())
                 .hobbyId(clubService.findHobbyId(clubRequest.getHobbyName()))
                 .maxMan(clubRequest.getMaxMan())
                 .build();
+        //게시물 작성 성공시 채팅방 생성 및 이미지 작성
         if (clubService.saveClub(club) == 1) {
             chattingService.openRoom(postId, principalDetails, "club");
-            List<ImageRequest> images = imageUtils.uploadImages(clubRequest.getImages());
-            clubService.saveImages(postId, images);
+            List<FileRequest> images = fileUtils.uploadFiles(clubRequest.getImages());
+            //이미지 업로드 성공시 DB에 이미지정보 저장
+            fileService.saveFiles(postId, images);
             model.addAttribute("result", "글 등록성공!");//나중에 삭제
         } else {
             model.addAttribute("result", "글 등록실패!");//나중에 삭제
@@ -108,6 +112,11 @@ public class ClubController {
     public String clubDetail(@RequestParam(value = "postId", required = false) Long postId, Model model,
                              HttpServletRequest request, HttpServletResponse response,
                              @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        ClubResponse response = clubService.findClub(postId, principalDetails);
+        List<FileResponse> files = response.getFileResponseList();
+        List<FileInfo> fileInfoList = fileUtils.getFileInfo(files);
+        model.addAttribute("images", fileInfoList);
+        model.addAttribute("club", response);
         ClubResponse clubResponse = clubService.findClub(postId, principalDetails);
         List<ImageResponse> imageResponses = clubResponse.getImageResponseList();
 
@@ -150,19 +159,69 @@ public class ClubController {
         return "club/clubDetail";
     }
 
+    @GetMapping("/club/update")
+    @ResponseBody
+    public ClubResponse update(@RequestParam(value = "postId") Long postId, Model model
+            , @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        //  로그인 안했을때
+        if (principalDetails == null) {
+//            return "redirect:clubDetail?postId=" + postId;
+            return null;
+        }
+        ClubResponse clubResponse = clubService.findClub(postId, principalDetails);
+        //  게시글 작성자가 아닐때
+        if (principalDetails.getMember().getMemberId() != clubResponse.getMemberId()) {
+//            return "redirect:clubDetail?postId=" + postId;
+            return null;
+        }
+        model.addAttribute("club", clubResponse);
+        return clubResponse;
+    }
+
+//    @PostMapping("/club/update")
+//    public String update(ClubRequest clubRequest, Model model
+//            , @AuthenticationPrincipal PrincipalDetails principalDetails) {
+//        //  로그인 안했을때
+//        if (principalDetails == null) {
+//            return "redirect:clubDetail?postId=" + postId;
+//        }
+//        ClubResponse clubResponse = clubService.findClub(postId, principalDetails);
+//        //  게시글 작성자가 아닐때
+//        if (principalDetails.getMember().getMemberId() != clubResponse.getMemberId()) {
+//            return "redirect:clubDetail?postId=" + postId;
+//        }
+//        clubService.updatePost()
+//    }
+
     @GetMapping("/clubRemove")
-    public String clubRemove(Long postId) {
+    public String clubRemove(@RequestParam(value = "postId") Long postId, @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        if (principalDetails == null) {
+            return "redirect:clubDetail?postId=" + postId;
+        }
+        //게시글삭제
         clubService.deletePost(postId);
+        //파일 리스트 가져오기
+        List<FileResponse> fileResponseList = fileService.findAllFileByPostId(postId);
+        //이미지 Disk 에서 삭제
+        fileUtils.deleteFiles(fileResponseList);
+        //이미지 DB 에서 삭제
+        fileService.deleteAllFileByIds(fileResponseList);
         return "redirect:clubList";
     }
 
-    // 파일 리스트 조회
-    @GetMapping("/posts/{postId}/images")
-    @ResponseBody
-    public List<ImageResponse> clubImage(@PathVariable Long postId) {
-        return clubService.findAllImageByPostId(postId);
-    }
+//    // 파일 리스트 조회
+//    @GetMapping("/posts/{postId}/images")
+//    @ResponseBody
+//    public List<FileResponse> clubImage(@PathVariable Long postId) {
+//        return clubService.findAllImageByPostId(postId);
+//    }
 
+    /**
+     * 모임가입
+     * @param postId 게시글아이디
+     * @param principalDetails 사용자 정보
+     * @return ResponseBody(Ajax)
+     */
     @PostMapping("/club/join")
     @ResponseBody
     public ClubResponse join(Long postId, @AuthenticationPrincipal PrincipalDetails principalDetails) {
